# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Organizations API",
        "description": "REST API для работы с организациями, зданиями и деятельностями",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API сервер"
        }
    ],
    "paths": {
        "/organizations": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Получить список организаций",
                "description": "Возвращает список организаций с возможностью фильтрации по зданию, названию, геопозиции и виду деятельности (включая дерево до 3 уровней).",
                "operationId": "3952cb0969bc8305f39882b2637f9b12",
                "parameters": [
                    {
                        "name": "building_id",
                        "in": "query",
                        "description": "ID здания для фильтрации",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Поиск по названию организации",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Широта для фильтрации по радиусу",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 54.7388
                        }
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "description": "Долгота для фильтрации по радиусу",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 55.9721
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "description": "Радиус поиска (в километрах)",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 1
                        }
                    },
                    {
                        "name": "activity_id",
                        "in": "query",
                        "description": "ID вида деятельности (учитывается дерево до 3 уровней)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "ООО Рога и Копыта"
                                            },
                                            "phones": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                },
                                                "example": [
                                                    "2-222-222",
                                                    "8-923-666-13-13"
                                                ]
                                            },
                                            "building": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "address": {
                                                        "type": "string",
                                                        "example": "г. Уфа, ул. Ленина 1"
                                                    },
                                                    "latitude": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 54.7388
                                                    },
                                                    "longitude": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 55.9721
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "activities": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Молочная продукция"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/organizations/{organization}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Получить организацию по ID",
                "operationId": "c0d4bab9fe7d2e1c7f2512135f9abb93",
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "description": "ID организации",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные организации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "123"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Название организации"
                                        },
                                        "building": {
                                            "$ref": "#/components/schemas/Building"
                                        },
                                        "activities": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Activity"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Организация не найдена"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Activity": {
                "title": "Activity",
                "description": "Модель деятельности",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Молочная продукция"
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Building": {
                "title": "Building",
                "description": "Модель здания",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "address": {
                        "type": "string",
                        "example": "г. Уфа, ул. Ленина 1"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 54.7388
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": 55.9721
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "API ключ для доступа к защищённым методам",
                "name": "x-api-key",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "ApiKeyAuth": []
        }
    ],
    "tags": [
        {
            "name": "Organizations",
            "description": "Organizations"
        }
    ]
}